ChatServer-master/ChatServer-master/requirements.txt
aioredis==1.3.0
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
boto3==1.10.26
botocore==1.13.26
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
Django==2.2.7
django-polymorphic==2.1.2
django-redis-sessions==0.6.1
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==1.4.6
phonenumbers==8.10.23
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
redis==3.3.11
s3transfer==0.2.1
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
zope.interface==4.7.1
